+++ R/sim_coor_sing-data.R
@@ -1,0 +1,18 @@
+#' Simulated coordinated singing events.
+#' 
+#' @description \code{sim_coor_sing} same data set as 'sim.coor.sing'. Selections of simulated interactive singing events.The simulated events use the mean 
+#' and standard deviation of real lekking \emph{Phaethornis longirostris} (Long-billed
+#' 
+#' Hermit hummingbird) songs and intervals between songs (e.i gaps). Three events are simulated: overlapping signals (ovlp), 
+#' alternating signals (altern) and non-synchronized signals (uncoor).
+#' 
+#' @format \describe{
+#'   
+#'   \item{sim_coor_sing}{Simulated coordinated singing events that overlap and do not overlap most of the time,
+#'   for use with \code{coor.test}}
+#'   
+#'   }
+#'   
+#' @usage data(sim_coor_sing)
+#' 
+"sim_coor_sing"